Fix budget phase label for 2013 outcome in 2015 budget
Jeffery Smith
08:11 (2 hours ago)
to me, Greg
Hi JD

I think it is just a heading error as we don’t collect preliminary outcomes for the history years. Prelim only relates to the in-year. I also double checked in the ENE data collection template, where we have it as audited outcome.

Jeffery